refactor(admin panel routes, forms): changes in response to client feedback on the project
changes added include, removal of navbar link to cohorts, removal of link to admin panel, changed to
the user list/table. Disabled user delete button, replaced with a deactive user button that is not
currently connected to the backend db. Edit button links to user to the same form; removed the Super
Users ability to assign them selves to as lecturers to papers. Bug that is still currently causing
issue, duplicate users have no error handling to prevent failure. Users must always be assigned to a
paper once they have been assigned to one. No option to not be assigned to a paper. Lecturer can be
made a super-user, but then not be able to self assign to other papers - this workflow needs to be
clarified.

#198, #172, #76, #209, #180