refactor(admin/lecturer permissions): lecturer role created to allow for them to be separated
Seeders changed provide an option for a new or previous lecturer to not be assigned to a paper,
removing the need to delete lecturers. removed the email from the user controller that was enforcing
the production super-admin role on a fake test site email account.

BREAKING CHANGE: evidence upload system is no longer functioning, tests in this branch, when run
locally do not upload files to the remote server for storage

#210, #128, #127