refactor(user admin forms): user Edit and create page ui updates + notes and evidence
page forms updated with tailwind css, improve layout. Forms used in place of a tag links for faster
response between page loads. Created new evidence migration for adding access key to evidence
uploads, key is hashed version of the file url - this may not be needed.

BREAKING CHANGE: form validations are not in place yet on the user forms, errors occur if incorrect
values are entered. this needs to be fixed before deployment.

#127, #128, #210, #205, #180, #93, #85, #76