fix(students.blade.php): working on debugging issues with student output
currently the inconsistent use of cohort_id and paper_id means that there is no functional
connection between students and papers that can be used to show students on a per paper basis on the
home page. Have managed to get to displaying the name of each paper the user is currently teaching,
but the list of students is duplicated for each due to different names used in each table. (i think,
this is the issue). I have attempted to scrap the search student component to use the component
added in the last sprint, this also has the same issue of not being able to access data.

BREAKING CHANGE: currently the inconsistent use of cohort_id and paper_id means that there is no
functional connection between students and papers that can be used to show students on a per paper
basis on the home page. Have managed to get to displaying the name of each paper the user is
currently teaching, but the list of students is duplicated for each due to different names used in
each table. (i think, this is the issue). I have attempted to scrap the search student component to
use the

#180, #181