fix(students.blade.php): bug fixes added to the student controller
issue came up with the way loops had been used to output cohorts, papers, students, and users.
errors resulted in the home page not working correctly. Some of the issues with this have been
fixed. I have created a new query in the student controller, which gets all the papers assigned to a
lecturer, matches the students in those papers, and returns them to the student page. the query
still needs to account for the cohort stream.

BREAKING CHANGE: Cohorts are not working correctly, I just need to alter the query in the student
controller to include this in the results

#180, #181, #206