fix(students.blade.php): lecturer class list now getting displayed
just need to update the student index query to get each students stream infomation if available.
cohort does not appear to have this now, so it may need to be added to the papers table, model -
will try to avoid doing that. The course selector drop-down is now also working, but is not
associating a student with the selected stream.

BREAKING CHANGE: paper selection dropdown is not currently associating a new student with a stream,
paper

#180, #181