fix(debugging getting students per lecturer): unable to get a connection between user and student
there are issues such as students can be assigned to multiple studio papers at the same time,
lecturers can have more than one paper, but the list of students is the same in all they are
assigned. There is a really complex web of tables, and I am struggling to make a connection with
them to show students in a paper taught by the currently logged in user.

BREAKING CHANGE: the front page cannot access data from the backend, the initial commit on this
branch was not working and is still not working due to the complex nature of the database

#180, #181