test(dusk config change + logout test script): completed dusk test for logout security
tests needed a small change to be make in the CreatesApplication file. Created a logout security
test that is successfully running. Made small change to a couple of components so these now have an
id attribute that can be used in the tests. Test script has function to run fresh migrations at the
start, then seeds the database. This script needs to be included as is across all tests to get past
the authentication.

#229, #221, #199, #112