fix(list style bug fixed): conditional statements used in student-list did not work
for non-admin users view. Removed these, so the original flaw is now part of the template.

BREAKING CHANGE: users with only one paper assigned have their list of students on the far right
side of the screen instead of being centered. Unable to work out how to access the paper count and
conditionally apply css based on if the user has only one cohort or more than one.

#179